refactor(chat): extract team access into hook
Move team-chat access + membership helpers (fetchTeamDetails,
revokeTeamChatAccess, refreshActiveTeamMembership,
hydrateTeamConversationDetails) and the 10s membership-polling effect
from Chat.jsx into src/hooks/useChatTeamAccess.js. Verbatim move;
fetchTeamDetails stays internal to the hook.
